Move project class under Project namespace
